revert to 1 comment at a time
