+++ src/commenter.ts
@@ -1,37 +1,7 @@
 /* eslint-disable @typescript-eslint/explicit-function-return-type */
 import * as core from '@actions/core'
 import * as github from '@actions/github'
-import {Octokit} from '@octokit/action'
-import {throttling} from '@octokit/plugin-throttling'
-
-const token = core.getInput('token') || process.env.GITHUB_TOKEN
-
-const ThrottlingOctokit = Octokit.plugin(throttling)
-const octokit = new ThrottlingOctokit({
-  auth: `token ${token}`,
-  throttle: {
-    onRateLimit: (
-      retryAfter: number,
-      options: any,
-      o: Octokit,
-      retryCount: number
-    ) => {
-      core.warning(
-        `Request quota exhausted for request ${options.method} ${options.url}
-Retry after: ${retryAfter} seconds
-Retry count: ${retryCount}
-`
-      )
-      return true
-    },
-    onSecondaryRateLimit: (retryAfter: number, options: any, o: Octokit) => {
-      core.warning(
-        `SecondaryRateLimit detected for request ${options.method} ${options.url}`
-      )
-      return true
-    }
-  }
-})
+import {octokit} from './octokit.js'
 
 const context = github.context
 const repo = context.repo
@@ -185,38 +155,34 @@ ${tag}`
       `Submitting review for PR #${pull_number}, total comments: ${this.reviewCommentsBuffer.length}`
     )
     try {
-      let batchNumber = 1
-      while (this.reviewCommentsBuffer.length > 0) {
-        const commentsBatch = this.reviewCommentsBuffer.splice(0, 30)
-        core.info(
-          `Posting batch #${batchNumber} with ${commentsBatch.length} comments`
-        )
+      for (const comment of this.reviewCommentsBuffer) {
+        core.info(`Posting comment: ${comment.message}`)
 
-        await octokit.pulls.createReview({
-          owner: repo.owner,
-          repo: repo.repo,
-          pull_number,
-          commit_id,
-          event: 'COMMENT',
-          comments: commentsBatch.map(comment => {
-            const commentData: any = {
-              path: comment.path,
-              body: comment.message,
-              line: comment.end_line,
-              start_side: 'RIGHT'
-            }
-
-            if (comment.start_line !== comment.end_line) {
-              commentData.start_line = comment.start_line
-            }
-
-            return commentData
+        if (comment.start_line !== comment.end_line) {
+          await octokit.pulls.createReviewComment({
+            owner: repo.owner,
+            repo: repo.repo,
+            pull_number,
+            commit_id,
+            body: comment.message,
+            path: comment.path,
+            line: comment.end_line,
+            start_side: 'RIGHT',
+            start_line: comment.start_line
           })
-        })
+        } else {
+          await octokit.pulls.createReviewComment({
+            owner: repo.owner,
+            repo: repo.repo,
+            pull_number,
+            commit_id,
+            body: comment.message,
+            path: comment.path,
+            line: comment.end_line
+          })
+        }
 
-        core.info(`Batch #${batchNumber} posted`)
-
-        batchNumber++
+        core.info(`Comment posted`)
       }
     } catch (e) {
       core.warning(`Failed to submit review: ${e}`)
