+++ src/review-comment.ts
@@ -1,7 +1,5 @@
 import * as core from '@actions/core'
 import * as github from '@actions/github'
-import {Octokit} from '@octokit/action'
-import {throttling} from '@octokit/plugin-throttling'
 import {Bot} from './bot.js'
 import {
   Commenter,
@@ -10,37 +8,9 @@ import {
   EXTRA_CONTENT_TAG,
   SUMMARIZE_TAG
 } from './commenter.js'
+import {octokit} from './octokit.js'
 import {Inputs, Options, Prompts} from './options.js'
 import * as tokenizer from './tokenizer.js'
-
-const token = core.getInput('token') || process.env.GITHUB_TOKEN
-
-const ThrottlingOctokit = Octokit.plugin(throttling)
-const octokit = new ThrottlingOctokit({
-  auth: `token ${token}`,
-  throttle: {
-    onRateLimit: (
-      retryAfter: number,
-      options: any,
-      o: Octokit,
-      retryCount: number
-    ) => {
-      core.warning(
-        `Request quota exhausted for request ${options.method} ${options.url}
-Retry after: ${retryAfter} seconds
-Retry count: ${retryCount}
-`
-      )
-      return true
-    },
-    onSecondaryRateLimit: (retryAfter: number, options: any, o: Octokit) => {
-      core.warning(
-        `SecondaryRateLimit detected for request ${options.method} ${options.url}`
-      )
-      return true
-    }
-  }
-})
 
 const context = github.context
 const repo = context.repo
